f-content-cards/home-promotion-card-2 - Clean up templates and tests
