+++ packages/f-content-cards/src/components/cardTemplates/homePromotionCard/_tests/HomePromotionCard1.test.js
@@ -1,19 +1,12 @@
 import { shallowMount } from '@vue/test-utils';
-import HomeContentCard from '../HomeContentCard.vue';
+import HomePromotionCard1 from '../HomePromotionCard1.vue';
 
 const testId = '__TEST_ID__';
-const ctaText = '__CTA_TEXT__';
-const url = '__URL__';
 
-const card = {
-    ctaText,
-    url
-};
-
-describe('contentCards › HomeContentCard', () => {
+describe('contentCards › HomePromotionCard1', () => {
     it('should apply the given test ID', () => {
         // Arrange & Act
-        const wrapper = shallowMount(HomeContentCard, {
+        const wrapper = shallowMount(HomePromotionCard1, {
             propsData: {
                 testId
             }
@@ -23,28 +16,12 @@ describe('contentCards › HomeContentCard', () => {
         expect(wrapper.find(`[data-test-id="${testId}"]`).exists()).toBe(true);
     });
 
-    it('should display a CTA link and label', () => {
-        // Arrange & Act
-        const wrapper = shallowMount(HomeContentCard, {
-            propsData: {
-                card,
-                testId
-            }
-        });
-
-        const cta = wrapper.find(`[data-test-id="${testId}--cta"]`);
-
-        // Assert
-        expect(cta.text()).toBe(ctaText);
-        expect(cta.attributes('href')).toBe(url);
-    });
-
     it('should apply given max-width to inner container', () => {
         // Arrange
         const containerMaxWidth = 1234;
 
         // Act
-        const wrapper = shallowMount(HomeContentCard, {
+        const wrapper = shallowMount(HomePromotionCard1, {
             propsData: {
                 testId,
                 containerMaxWidth
